Move files out of new module
