+++ fcp-zbus/src/candidate.rs
