--- fcp-zbus/src/new/mod.rs
@@ -1,10 +1,0 @@
-pub mod pipeline;
-mod mode_switcher;
-mod dispatcher;
-mod cloud_pinyin_client;
-mod candidate_service;
-mod candidate;
-mod symbol_service;
-mod number_service;
-mod ibus_proxy;
-pub mod ibus_variants;
